requirements.txt
delta-spark
pandas
pyarrow
virtualenv